segfault-moderation/requirements.txt
azure-functions
pillow
requests
google-genai
psycopg2-binary
